pyproject.toml
[tool.poetry]
name = "headhunterparser"
version = "0.1.0"
description = ""
authors = ["Constantine Putilov <putilov.constantine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
psycopg2 = "^2.9.9"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
